Make the first experimental test work (#12)
In this particular case, it is sufficient to wait for the first text
change in the `<main>` element. There will be cases where this is not
sufficient, i.e., when nested or repeated views are to be rendered.
So we will probably be stuck with custom per-case waiting
instructions. I am afraid that this is inherent to the setup with
Backbone.