Rework revision creation logic for better attribution (#29)
A separate import revision is now created for each question group.
This enables attribution of a single author per group version. When
multiple authors are known, this is mentioned in the commit message.

Other solutions are possible, such as attaching multiple authors to
each revision. However, it is too early to justify changes to the
data model for this purpose.