Attempt to debug a logic error that defies all common sense
For some reason, control flow is entering the `else` branch despite
`authors` being `[None]`. As a result, `None` is passed as the `name`
to `get_person`, which leads to the following exception:

AttributeError: 'NoneType' object has no attribute 'strip'

I'm a bit lost now, as the Python interpreter seems to be cheating on
the language semantics. Probably best to take a break.

ref: #29